Make `HttpSwikiRequest>>#raw` redirect to `self`
In Swiki there are some features that coupling with `HttpRequest` of
`KomHttpServer`, via `HttpSwikiRequest>>#raw`. Now Swiki have been
migrated to `WebServer`, there need to be a way to simulate the
interface of `HttpRequest`, which will be the responsibility of
`HttpSwikiRequest`.

This commit makes `HttpSwikiRequest>>#raw` redirect to `self`, and
further compatibility features will be implemented in `HttpSwikiRequest`
directly, also, in this way, the new `raw` (i.e. `WebRequest`) will not
be exposed any more.